requirements.txt
streamlit 
pandas 
google-genai
plotly 
